refactor(LLM): remove unnecessary hook and fix old bugs
