+++ .changeset/tricky-eels-remain.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": minor
+---
+
+Refactor AuthPath
