+++ apps/ledger-live-mobile/__tests__/test-renderer.tsx
@@ -40,6 +40,7 @@ import { INITIAL_STATE as TOASTS_INITIAL_STATE } from "~/reducers/toast";
 import { State } from "~/reducers/types";
 import { INITIAL_STATE as WALLET_CONNECT_INITIAL_STATE } from "~/reducers/walletconnect";
 import { INITIAL_STATE as WALLETSYNC_INITIAL_STATE } from "~/reducers/walletSync";
+import { INITIAL_STATE as AUTH_INITIAL_STATE } from "~/reducers/auth";
 import StyleProvider from "~/StyleProvider";
 import CustomLiveAppProvider from "./CustomLiveAppProvider";
 import { getFeature } from "./featureFlags";
@@ -66,6 +67,7 @@ const INITIAL_STATE: State = {
   wallet: WALLET_INITIAL_STATE,
   walletconnect: WALLET_CONNECT_INITIAL_STATE,
   walletSync: WALLETSYNC_INITIAL_STATE,
+  auth: AUTH_INITIAL_STATE,
 };
 
 type ExtraOptions = RenderOptions & {
